+++ bazel/src/main/java/com/google/api/codegen/bazel/BazelBuildFileView.java
@@ -16,6 +16,7 @@ package com.google.api.codegen.bazel;
 
 import java.lang.String;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -44,7 +45,6 @@ class BazelBuildFileView {
     tokens.put("version", bp.getVersion());
     tokens.put("package", bp.getProtoPackage());
     tokens.put("migration_mode", '"' + bp.getPhpMigrationMode() + '"');
-
 
     // For regeneration of Java rules, we are particularly interested in what the saved transport value was,
     // if there was one, in order to correctly generate, or not, the rest/grpc specific targets and labels.
@@ -193,8 +193,10 @@ class BazelBuildFileView {
     // Construct GAPIC import path & package name based on go_package proto option
     String protoPkg = bp.getProtoPackage();
     boolean isCloud = bp.getCloudScope() || protoPkg.contains("cloud");
-    String goImport = assembleGoImportPath(isCloud, protoPkg, bp.getLangProtoPackages().get("go"));
-
+    String goImport = bp.getGoImportpathOverride();
+    if (goImport == null) {
+      goImport = assembleGoImportPath(isCloud, protoPkg, bp.getLangProtoPackages().get("go"));
+    }
     tokens.put("go_gapic_importpath", goImport);
     tokens.put("go_gapic_test_importpath", goImport.split(";")[0]);
     tokens.put("go_gapic_deps", joinSetWithIndentationNl(mapGoGapicDeps(actualImports)));
@@ -211,6 +213,18 @@ class BazelBuildFileView {
     // Ideally, we'd use a slightly more sophisticated templating system, like Mustache, that would
     // allow us to omit `rest_numeric_enums` when `!transport.contains("rest")`
     tokens.put("rest_numeric_enums", numericEnums);
+
+    // Again, ideally we wouldn't emit extra_opts for C# at all unless we had some...
+    Map<String, List<String>> csharpListAttributes = bp.getOverriddenListAttributes().get("csharp");
+    if (csharpListAttributes != null) {
+      List<String> csharpProtoExtraOpts = csharpListAttributes.get("extra_opts");    
+      if (csharpProtoExtraOpts != null) {
+        tokens.put("csharp_proto_extra_opts", joinCollectionWithIndentation(csharpProtoExtraOpts));
+      }
+    }
+    if (!tokens.containsKey("csharp_proto_extra_opts")) {
+      tokens.put("csharp_proto_extra_opts", "");
+    }
   }
 
   /**
@@ -297,7 +311,11 @@ class BazelBuildFileView {
   }
 
   private String joinSetWithIndentation(Set<String> set) {
-    return set.isEmpty() ? "" : '"' + String.join("\",\n        \"", set) + "\",";
+    return joinCollectionWithIndentation(set);
+  }
+
+  private String joinCollectionWithIndentation(Collection<String> collection) {
+    return collection.isEmpty() ? "" : '"' + String.join("\",\n        \"", collection) + "\",";
   }
 
   private String joinSetWithIndentationNl(Set<String> set) {